Got the calc_bkgd program up and running. - Fixed some problems with the reading of HapMap rec rates. - Added a function to validate the recombination map. - Made a config file with extensive comments.

+++ bio/seqfeat.h
@@ -29,7 +29,7 @@ struct SeqFeature_struct {
   SeqCoord c;  /* genomic coordinate */
   long id;     /* database identifier */
   char *name;  /* name of feature */
-  float score; /* score of feature */
+  double score; /* score of feature */
 
   SeqFeatAttrib *attrib; /* ptr to list of other named attributes */
